Fix bad ctr flag and scope-agent clippy errors
containerd/mod.rs: remove --memory-swap-limit (not a valid ctr run flag);
instead disable swap post-start by writing 0 to the container's cgroupv2
memory.swap.max file via the pid returned by run_container

scope/agent: fix four clippy -D warnings errors:
- enrichment.rs: derive Default on EnrichmentData instead of manual impl,
  add impl Default for EnrichmentProvider delegating to new()
- otel.rs: remove redundant i64->i64 cast on syscall_nr
- probes.rs: add impl Default for ProbeManager delegating to new()

+++ scope/agent/src/otel.rs
@@ -191,7 +191,7 @@ impl OtelExporter {
         enrichment: Option<&EnrichmentData>,
     ) {
         let mut attributes = vec![
-            KeyValue::new("syscall.nr", sys_event.syscall_nr as i64),
+            KeyValue::new("syscall.nr", sys_event.syscall_nr),
             KeyValue::new("event.type", format!("{:?}", sys_event.event_type)),
             KeyValue::new("process.pid", _event.pid as i64),
             KeyValue::new("process.comm", _event.comm.clone()),
